Importance of test automation in industry
Automated testing is the best solution to overcome the problems faced in manual testing method. Automated testing tools are able to playback the pre-defined and pre-recorded actions, comparing the results with the expected behavior of the application and then report all the results to the test engineers. Once all of the automated tests have been created, then they can be easily repeated. They can also be extended so they can perform the tasks that are impossible to do with manual testing method.
It has been a long time since automation testing is considered to be a very critical step for big organizations that develops applications. But small companies cannot afford to implement because it can be costly solution to have. They need an appropriate and affordable solution in order to complete their projects successfully with time & money spare. With automated testing, the repetitive tasks are done in a really shot time as compared to the manual testing method. The automated testing can easily execute many different test cases. And during each test, it provides coverage which is really not possible in the manual testing method.